+++ arm9/source/romBrowser/RomBrowserStateTrigger.h
@@ -14,6 +14,8 @@ enum class RomBrowserStateTrigger
     GotoSettingsScreen,
     ShowRecents,
     HideRecents,
+    ShowFavorites,
+    HideFavorites,
     ShowStatistics,
     HideStatistics,
     ShowDeleteConfirm,