+++ arm9/source/romBrowser/viewModels/RomBrowserAppBarViewModel.h
@@ -43,6 +43,11 @@ public:
         _romBrowserController->ShowRecents();
     }
 
+    void ShowFavorites()
+    {
+        _romBrowserController->ShowFavorites();
+    }
+
     void RequestDeleteSelected()
     {
         _romBrowserController->RequestDeleteSelected();